+++ CHANGE_LOG.md
@@ -1,7 +1,9 @@
-# v5.8.4
+# v5.8.5
 - Distuiguish between oid and sub claims for license
 - Make Validation Key a string
 - Deployment check for license
+- Added issued_at and expiry_at in LicenseFile model
+- Remove secondary key and simplify licensing
 
 # v5.8.1
 - Rename auth_apps claim to apps